Improvements to golang-single class.
Treat gopkg.in/ as a special import path. This way is possible to roll
out the entire chain of imported paths without the need to rely on
workarounds based on function golang_fix_importpath_alias().

In the next days, golang.x/ and several other paths will be added
as special import paths too. Cleanup of ebuilds will follow ASAP.